+++ pages/teach-like-mozilla.jsx
@@ -13,9 +13,7 @@ var TeachLikeMozillaPage = React.createClass({
   render: function() {
     return (
       <div>
-        <HeroUnit
-          image="/img/pages/teach-like-mozilla/hero-teach-like-mozilla.png"
-          image2x="/img/pages/teach-like-mozilla/hero-teach-like-mozilla@2x.png">
+        <HeroUnit>
           <h1>Teach Like Mozilla</h1>
           <h2>We learn best by making & reflecting, together.</h2>
         </HeroUnit>
@@ -53,21 +51,18 @@ var TeachLikeMozillaPage = React.createClass({
               <IconLink
                 href="http://mzl.la/TTWpodcasts"
                 imgSrc="/img/pages/teach-like-mozilla/svg/icon-listen.svg"
-                imgAlt="icon listen"
                 head="Listen"
                 subhead="Subscribe to our podcast"
               />
               <IconLink
                 linkTo="web-literacy"
                 imgSrc="/img/pages/teach-like-mozilla/svg/icon-learn.svg"
-                imgAlt="icon learn"
-                head="Learn"
-                subhead="See our Web Literacy Map"
+                head="Understand"
+                subhead="Learn more about the Web Literacy Map"
               />
               <IconLink
                 href="http://discourse.webmaker.org/category/meet"
                 imgSrc="/img/pages/teach-like-mozilla/svg/icon-connect.svg"
-                imgAlt="icon connect"
                 head="Say Hello"
                 subhead="Meet the teach community"
               />