+++ pages/web-literacy.jsx
@@ -189,21 +189,18 @@ var ActivitiesPage = React.createClass({
             <IconLink
               href="https://wiki.mozilla.org/Webmaker/WebLiteracyMap"
               imgSrc="/img/pages/web-literacy/svg/icon-contribute.svg"
-              imgAlt="icon contribute"
               head="Contribute"
               subhead="Join the community on defining Web Literacy at Mozilla"
             />
             <IconLink
               href="https://mozilla.github.io/webmaker-whitepaper/"
               imgSrc="/img/pages/web-literacy/svg/icon-learn-more.svg"
-              imgAlt="icon learn more"
               head="Learn More"
               subhead="This whitepaper explores why Mozilla cares about Web Literacy"
             />
             <IconLink
               href="https://www.mozilla.org/en-US/about/manifesto/"
               imgSrc="/img/pages/web-literacy/svg/icon-reading.svg"
-              imgAlt="icon reading"
               head="Get Reading"
               subhead="Web Literacy is core to the Mozilla mission and values"
             />