+++ CHANGELOG.md
@@ -6,6 +6,94 @@ to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
 
+## [0.0.9][] - 2015-06-03
+### Added
+- A new "Protect Your Data" teaching activity has been added,
+  and is now available at `activities/protect-your-data/`.
+
+### Changed
+- New clubs are no longer automatically publicly visible; instead, they
+  are initially set to a "pending" state, and can be approved or
+  denied by staff on the back-end. A "My Clubs" section above the
+  rest of the clubs now informs the user about their clubs and their
+  approval statuses. Users are also informed about the approval flow
+  when they add a new club.
+- The list of recent blog posts on the homepage is no longer hard-coded,
+  but dynamically pulled from the blog's RSS feed via the Google Feed API.
+  (We thought this was in 0.0.8, but it wasn't.)
+- The "Learn" CTA at the bottom of the Teach Like Mozilla page
+  has been changed to "Understand", with the sub-copy changed to
+  "Learn more about the Web Literacy Map".
+- The Home and About pages now consistently refer to "the Mozilla
+  Learning Network", rather than "Mozilla Learning Networks".
+- The decorative icons at the bottom CTAs of pages now have null
+  `alt` text, improving accessibility.
+- When sharing a Teach site page on Facebook, the page description is
+  now the blurb from the About Us page ("We want more people to see
+  themselves as citizens of the web...") rather than the footer text
+  ("The Hive Learning Networks, stewarded by Mozilla...").
+- When the Web Literacy Basics page is shown, the Mozilla Clubs section of
+  the nav is no longer expanded.
+
+## [0.0.8][] - 2015-06-01
+### Added
+- The add club modal now requires users to mark a checkbox labeled
+  "I have read the Mozilla Clubs Fact Sheet".
+
+### Changed
+- ~~The list of recent blog posts on the homepage is no longer hard-coded,
+  but dynamically pulled from the blog's RSS feed via the Google Feed API.~~
+  (This change didn't actually make it into the release.)
+- The add/change club modal has better prompts so content featured
+  publicly feels more in fidelity with clubs.
+- Pages now load faster on devices with retina displays.
+
+## [0.0.7][] - 2015-05-28
+### Added
+- `srcset` attributes are now link-checked during automated testing.
+- The dev version modal now *always* links to the latest release
+  of the site, and potentially a link to the changes between the
+  latest release and the current commit. It also links to
+  `CONTRIBUTING.md`.
+- A reference to novalidate.com has been added to `CONTRIBUTING.md`.
+
+### Changed
+- Homepage has been refreshed to better inform
+  the site vistors of what they can do on the site.
+- Homepage now shows the latest blog posts from
+  https://blog.webmaker.org/.
+- Various performance improvements to the manual test page have
+  been added.
+
+## [0.0.6][] - 2015-05-26
+### Added
+- A number of accessibility improvements have been added:
+  - Pages on the site now have `contentinfo`, `main`, and
+    `navigation` landmark roles.
+  - Pages now indicate that they are in English.
+  - More purely decorative images have null `alt` attributes.
+  - Skip navigation has been added at the top of the page.
+  - Expander components on the Clubs Toolkit and Web Literacy pages now
+    use heading tags instead of divs.
+  - The logo assets in the Event Resources page are now keyboard
+    accessible.
+
+### Changed
+- All mentions of the @Webmaker Twitter handle (in the footer and
+  the About and Clubs Toolkit pages) have been changed to @MozTeach.
+- The `dist` dir is now cleaned up frequently during
+  development, ensuring that old files don't stay
+  around for too long.
+- `lib/changelog.js bump` now makes it easier to issue
+  new releases of the site.
+- The dev version modal now links to the latest release
+  of the site if the current git revision is unavailable.
+
+### Removed
+- `gulp-html-prettify` is no longer listed as a dependency
+  in `package.json` (it hasn't been used by any code for
+  some time).
+
 ## [0.0.5][] - 2015-05-20
 ### Added
 - A link to "Site Administration" now appears under the logout
@@ -21,7 +109,7 @@ to [Semantic Versioning](http://semver.org/).
 - Added environment variables `MAILINGLIST_URL`
   and `MAILINGLIST_PRIVACY_NAME`.
 - The site now generates `index.html` files for redirects. This
-  was done so that `/clubs/curriculum/` redirects to 
+  was done so that `/clubs/curriculum/` redirects to
   `/activities/web-lit-basics/`.
 - `lib/changelog.js` can be used as a node script to help
   update `CHANGELOG.md`.
@@ -29,7 +117,7 @@ to [Semantic Versioning](http://semver.org/).
   changelog are now verified by the automated test suite.
 
 ### Changed
-- The Clubs Curriculum page has been renamed to 
+- The Clubs Curriculum page has been renamed to
   Web Literacy Basics, and is now available at
   `/activities/web-lit-basics/`. It's no longer accessible
   through the sidebar, but is instead listed as the
@@ -43,7 +131,7 @@ to [Semantic Versioning](http://semver.org/).
   collapsed area becomes focused, improving navigation for
   sighted users who can only use the keyboard for navigation.
 - Signing up for the Maker Party mailing list now shows
-  a "Thank you" modal when the user is returned to the 
+  a "Thank you" modal when the user is returned to the
   site from BSD.
 - Moved all documentation about manual testing to
   `CONTRIBUTING.md`.
@@ -117,7 +205,11 @@ to [Semantic Versioning](http://semver.org/).
 ### Added
 - This is the initial release, pushed to https://teach.mozilla.org/.
 
-[unreleased]: https://github.com/mozilla/teach.webmaker.org/compare/v0.0.5...HEAD
+[unreleased]: https://github.com/mozilla/teach.webmaker.org/compare/v0.0.9...HEAD
+[0.0.9]: https://github.com/mozilla/teach.webmaker.org/compare/v0.0.8...v0.0.9
+[0.0.8]: https://github.com/mozilla/teach.webmaker.org/compare/v0.0.7...v0.0.8
+[0.0.7]: https://github.com/mozilla/teach.webmaker.org/compare/v0.0.6...v0.0.7
+[0.0.6]: https://github.com/mozilla/teach.webmaker.org/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/mozilla/teach.webmaker.org/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/mozilla/teach.webmaker.org/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/mozilla/teach.webmaker.org/compare/v0.0.2...v0.0.3